move analytics service to core bundle, make sotrage bundle work with symfony4
