+++ src/SWP/Bundle/AnalyticsBundle/EventListener/AnalyticsEventListener.php
@@ -53,13 +53,13 @@ class AnalyticsEventListener
     public function onKernelRequest(GetResponseEvent $event)
     {
         $request = $event->getRequest();
-        if (null !== ($json = file_get_contents('php://input')) && '' !== $json) {
-            $request->attributes->set('data', \json_decode($json, true));
-        } elseif (null !== $json = $request->getContent()) {
-            $request->attributes->set('data', \json_decode($json, true));
-        }
-
         if (strpos($request->getPathInfo(), self::EVENT_ENDPOINT)) {
+            if (null !== ($json = file_get_contents('php://input')) && '' !== $json) {
+                $request->attributes->set('data', \json_decode($json, true));
+            } elseif (null !== $json = $request->getContent()) {
+                $request->attributes->set('data', \json_decode($json, true));
+            }
+
             $this->producer->publish(serialize($request));
 
             $response = new Response();
