+++ src/SWP/Bundle/AnalyticsBundle/Resources/config/services.yml
@@ -8,12 +8,3 @@ services:
             - { name: kernel.event_listener, event: kernel.response, method: onKernelResponse, priority: 20 }
             - { name: kernel.event_listener, event: kernel.finish_request, method: onKernelFinishRequest, priority: 1 }
             - { name: kernel.event_listener, event: kernel.terminate, method: onKernelTerminate, priority: 1 }
-
-    SWP\Bundle\AnalyticsBundle\Services\ArticleStatisticsService:
-        public: true
-        arguments:
-            - '@swp.repository.article'
-            - '@swp.repository.article_statistics'
-            - '@swp.repository.article_events'
-            - '@swp.factory.article_statistics'
-            - '@swp.factory.article_events'
