+++ src/SWP/Bundle/AnalyticsBundle/composer.json
@@ -12,7 +12,8 @@
     }
   ],
   "require": {
-    "php": "^7.1"
+    "php": "^7.1",
+    "php-amqplib/rabbitmq-bundle": "^1.10"
   },
   "require-dev": {
     "phpunit/phpunit": "5.*"
