+++ src/SWP/Bundle/CoreBundle/Resources/config/services.yml
@@ -743,3 +743,12 @@ services:
         public: true
         arguments:
             - '@router'
+
+    SWP\Bundle\CoreBundle\Services\ArticleStatisticsService:
+        public: true
+        arguments:
+            - '@swp.repository.article'
+            - '@swp.repository.article_statistics'
+            - '@swp.repository.article_events'
+            - '@swp.factory.article_statistics'
+            - '@swp.factory.article_events'
