+++ src/SWP/Bundle/StorageBundle/composer.json
@@ -20,8 +20,8 @@
     "require": {
         "php": "^7.0",
         "swp/storage": "^1.4",
-        "symfony/framework-bundle": "^3.1",
-        "symfony/doctrine-bridge": "^3.1",
+        "symfony/framework-bundle": "^3.1|^4.0",
+        "symfony/doctrine-bridge": "^3.1|^4.0",
         "doctrine/orm": "^2.5",
         "doctrine/doctrine-bundle": ">=1.6.7 <1.7.0"
     },
